+++ extensions/iceworks-project-creator/web/src/pages/CreateProject/index.tsx
@@ -184,7 +184,7 @@ const CreateProject: React.FC = () => {
         ...curProjectField,
         clientToken: CLIENT_TOKEN,
       });
-      await callService('common', 'saveDataToSettingJson', 'user', { empId, account, gitlabToken });
+      await callService('common', 'saveUserInfo', { empId, account, gitlabToken });
       await callService('common', 'saveDataToSettingJson', 'workspace', projectPath);
       await callService('project', 'openLocalProjectFolder', projectDir);
     } catch (e) {