+++ extensions/iceworks-time-master/src/views/timerStatusBar.ts
@@ -1,10 +1,13 @@
-import { window, StatusBarAlignment, StatusBarItem } from 'vscode';
+import { window, StatusBarAlignment, StatusBarItem, workspace, ConfigurationChangeEvent } from 'vscode';
 import { getUserSummary } from '../storages/user';
 import { humanizeMinutes, seconds2minutes } from '../utils/time';
 import { getAverageSummary } from '../storages/average';
+import { CONFIG_KEY_ICEWORKS_ENABLE_STATUS_BAR, CONFIG_KEY_SECTION_ENABLE_STATUS_BAR } from '../constants';
+import { getDataFromSettingJson } from '@iceworks/common-service';
 
 interface TimerStatusBar extends StatusBarItem {
   refresh(): Promise<void>;
+  activate(): void;
 }
 
 export async function createTimerStatusBar() {
@@ -17,6 +20,23 @@ export async function createTimerStatusBar() {
   statusBar.command = 'iceworks-time-master.displayTimerTree';
   statusBar.refresh = async function () {
     statusBar.text = await getStatusBarText();
+  };
+  statusBar.activate = function () {
+    const enableStatusBar = getDataFromSettingJson(CONFIG_KEY_SECTION_ENABLE_STATUS_BAR);
+    if (enableStatusBar) {
+      statusBar.show();
+    }
+    workspace.onDidChangeConfiguration((event: ConfigurationChangeEvent) => {
+      const isChanged = event.affectsConfiguration(CONFIG_KEY_ICEWORKS_ENABLE_STATUS_BAR);
+      if (isChanged) {
+        const newEnableStatusBar = getDataFromSettingJson(CONFIG_KEY_SECTION_ENABLE_STATUS_BAR);
+        if (newEnableStatusBar) {
+          statusBar.show();
+        } else {
+          statusBar.hide();
+        }
+      }
+    });
   };
   return statusBar;
 }