+++ extensions/iceworks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.16.1
+
+- feat: settings for control Time Master Views
+- chore: update O2 toolbar button text
+- fix: do not auto set material source after user manually updates the settings
+- refactor: save user info to iceworks configuration
+- refactor: disable some duplicate features(extensions) in O2
+
 ## 0.16.0
 
 - feat: Project Dashboard